Implemented exit builtin
exit will only exit the shell if it was the only command token.
Redirects still happen, even heredoc is read, though they do nothing.
Thus we check if there is a next cmd node or not, and do it after
heredoc parsing.

Should implement other redirects too if we want bash behaviour, but it's
stupid, even to do it after heredoc and would like to check before
trying to parse them.

Since we are required to free everything before exiting the program
(even reachables aren't fine I guess?), ft_run_cmds is made to return an
int to let function interactive know if it should stop looping. Other
possibility is to give more arguments to ft_run_cmds, or probably better
is to have more attributes in the t_minishell struct.