Support for non-exported variables
Similar check to exit builtin, only make a var if it is the only token
and has a '=' character in it. This means only one variable at a time.
Bash can do multiple and also makes a temporary exported env char ** if
it were coupled with a command after it.