Fixed the exit builtin a ton
Forgot to allow for arguments to the exit cmd. Bash does a check to see
if it doesn't have too many arguments, it can only take one. Had to make
a new ft_atoi that checks for overflow over long and doesn't allow for
any non-digit characters except for a single leading sign.

I've also changed the t_minishell struct to include the line, cmd_list
and token_list to make it a bit easier to call exit_handler. This also
meant that the previous loop check in 'interactive' is not needed and
reverted the ft_run_cmds to a void return.