Support for temporary variables.
I recognized that the variable put in front of a command are just able
to be made in the child process replicating the behaviour from bash.
I've implemented this by checking the first element for the cmd_av to
see if it is a variable assignment. As long as this is the case, the
variable is made and the cmd_av is updated to not include that token.

Small change to let 'env_add_back' return with an int to save a line in
'env_add_var' so that the 'is_exp' can be set when the variable exists
already.

Next up is to implement the checks for the naming rules of the
variables. Currently it can include some characters that shouldn't be
allowed.