Small fix in parser and leak after syntax error
Parser assigned the REDIRECT tag if the token matched 'END'. This isn't
needed and would result in falsly seeing a heredoc '<< END' as a syntax
error.

After a syntax error regarding the redirects, the line was freed, but
not the tokens. Now it is, to fix a leak if you entered a line after
having a syntax error.