Moved two functions to cmd_special_checks.c
check_for_only_vars to see if the input only asks to make variable so
and adds them if that is the case. Then output is used to skip going
into a child process.

check_for_leading_vars wants to add variables to the environment while
in a child process for the variables that come before an actual command
token. Needs to update the cmd_av.